Revert "Refactor banner upload"
