+++ app/src/components/VendorProfile/VendorProfile.js
@@ -12,10 +12,9 @@ import banner from "../../styles/css/vendor_banner.module.css";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSave, faPen, faUpload } from "@fortawesome/free-solid-svg-icons";
 
-import BannerUploader from './BannerUploader';
 import { Image, CloudinaryContext, Transformation } from "cloudinary-react";
-import axiosWithAuth from "../../utils/axiosWithAuth";
-
+import axios from "axios";
+import axiosWithAuth from '../../utils/axiosWithAuth';
 
 const VendorProfile = props => {
   const [modal, setModal] = useState(false);
@@ -38,6 +37,8 @@ const VendorProfile = props => {
   const vendorId = props.match.params.id;
   const [editAbout, setEditAbout] = useState(false);
   const [editBusinessName, setEditBusinessName] = useState(false);
+
+
 
   const myWidget = window.cloudinary.createUploadWidget(
     {
@@ -89,10 +90,15 @@ const VendorProfile = props => {
       axiosWithAuth().put(
         `https://quickstlabs.herokuapp.com/api/v1.0/vendors/${vendorId}`,
 
+        { ...vendorInfo, vendor_banner: `${bannerInfo}` }
+      ).then(console.log('PUT request Issue'));
+    }
+  );
+
   useEffect(() => {
     async function fetchVendorInfo() {
       try {
-        const vendorInfo = await axiosWithAuth().get(
+        const vendorInfo = await axios.get(
           `https://quickstlabs.herokuapp.com/api/v1.0/vendors/${vendorId}`
         );
         console.log(`vendorinfo changed`, vendorInfo);
@@ -106,7 +112,7 @@ const VendorProfile = props => {
 
     async function fetchProducts() {
       try {
-        const products = await axiosWithAuth().get(
+        const products = await axios.get(
           `https://quickstlabs.herokuapp.com/api/v1.0/vendors/${vendorId}/products`
         );
 
@@ -125,7 +131,7 @@ const VendorProfile = props => {
       if (productIds.length !== 0) {
         for (const ids of productIds) {
           try {
-            const imageIds = await axiosWithAuth().get(
+            const imageIds = await axios.get(
               `https://quickstlabs.herokuapp.com/api/v1.0/products/${ids}/product-images`
             );
 
@@ -194,6 +200,11 @@ const VendorProfile = props => {
       });
   };
 
+  const uploadBanner = e => {
+    e.preventDefault();
+    myWidget.open();
+  };
+
 
   /* setShow(!show) */
   const reveal = () => {
@@ -202,6 +213,7 @@ const VendorProfile = props => {
   }
 
   console.log('dropdown_links', document.getElementById(`${profile.dropdown_links}`))
+
   return (
     <div className={profile.vendor_profile_container}>
       <div className={profile.vendor_header_container}>
@@ -289,21 +301,24 @@ const VendorProfile = props => {
             </CloudinaryContext>
           ) : (
               <img
-                className="vendor_banner_image"
+                className={banner.vendor_banner_image}
                 src={picture}
                 alt="vendor header"
               />
             )}
           <div className={banner.vendor_banner_upload}>
-            <BannerUploader
-              vendorId={vendorId}
-              vendorInfo={vendorInfo}
-              setBannerInfo={setBannerInfo}
-              bannerInfo={bannerInfo}
-
-
+            <FontAwesomeIcon
+              id={banner.upload}
+              className={banner.icon}
+              icon={faUpload}
+              onClick={uploadBanner}
             />
-          </div>
+            {/* <img
+              src={upload}
+              alt='upload icon'
+              onClick={uploadBanner}
+            /> */}
+          </div>{" "}
         </div>
       </div>
